+++ Sources/Sentry/include/SentryPrivate.h
@@ -1,3 +1,5 @@
 // Sentry internal headers that are needed for swift code
 
 #import "SentryBaseIntegration.h"
+#import "SentryRandom.h"
+#import "SentryTime.h"