+++ Tests/SentryTests/Integrations/Performance/AppStartTracking/SentryAppStartTrackerTests.swift
@@ -1,4 +1,5 @@
 import Nimble
+@testable import Sentry
 import SentryTestUtils
 import XCTest