+++ Tests/SentryTests/Integrations/WatchdogTerminations/SentryWatchdogTerminationsTrackerTests.swift
@@ -1,3 +1,4 @@
+@testable import Sentry
 import SentryTestUtils
 import XCTest