+++ Tests/SentryTests/SentrySDKIntegrationTestsBase.swift
@@ -1,4 +1,5 @@
 import Foundation
+@testable import Sentry
 import SentryTestUtils
 import XCTest